:recycle: refactor plots
